+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.4.0 - 2025-11-11
 
-* support --noempty and --cache cli option
+* support --noempty, --cache and --symlinks cli option
 * added preferences screen
 
 ## 0.3.2 - 2025-11-11